requirements.txt
hvac==0.9.6
openshift==0.9.0
ruamel.yaml==0.15.97
requests_mock
parameterized==0.7.4